+++ MOP/src/GameObjects/Vehicles/Satsuma.cs
@@ -48,9 +48,12 @@ namespace MOP
 
             for (int i = 0; i < disableableObjects.Length; i++)
             {
+                if (disableableObjects[i] == null)
+                    continue;
+
                 disableableObjects[i].gameObject.SetActive(enabled);
             }
-        } 
+        }
 
         /// <summary>
         /// Get list of disableable childs.