+++ MOP/src/Misc/MopSettings.cs
@@ -14,7 +14,9 @@
         public static float ActiveDistanceMultiplicationValue { get; private set; }
 
         public static bool SafeMode { get; set; }
+
         public static bool ToggleVehicles { get; private set; }
+
         public static bool ToggleItems { get; private set; }
 
         //
@@ -40,42 +42,55 @@
         static int occlusionSampleDelay = 1;
         public static int OcclusionSampleDelay { get => occlusionSampleDelay; }
 
-        // OcclusionHideDelay is calculated automatically
-        public static int OcclusionHideDelay = -1;
-
         static int minOcclusionDistance = 50;
         public static int MinOcclusionDistance { get => minOcclusionDistance; }
 
         static OcclusionMethods occlusionMethod = OcclusionMethods.Chequered;
         public static OcclusionMethods OcclusionMethod { get => occlusionMethod; }
 
-        // others
-        static bool warningShowed;
-        public const int UnityCarActiveDistance = 5; 
+        //
+        // MISCELLANEOUS
+        //
+
+        // OcclusionHideDelay is calculated automatically
+        public static int OcclusionHideDelay = -1;
+
+        // Checked after the warning about maximum distance and minimum distance of occlusion culling is showed,
+        // so it won't be displayed again.
+        static bool occlusionDistanceWarningShowed;
+
+        // Distance after which car physics are toggled
+        public const int UnityCarActiveDistance = 5;
+
+        // If true, vehicles physics won't be disabled on small distances, but only on larger ones
+        static bool overridePhysicsToggling = false;
+        public static bool OverridePhysicsToggling { get => overridePhysicsToggling; }
 
         public static void UpdateAll()
         {
+            // Activating Objects
             ActiveDistance = int.Parse(MOP.activeDistance.GetValue().ToString());
             ActiveDistanceMultiplicationValue = GetActiveDistanceMultiplicationValue();
-            
             SafeMode = (bool)MOP.safeMode.GetValue();
             ToggleVehicles = (bool)MOP.toggleVehicles.GetValue();
             ToggleItems = (bool)MOP.toggleItems.GetValue();
 
+            // Others
             removeEmptyBeerBottles = (bool)MOP.removeEmptyBeerBottles.GetValue();
             satsumaTogglePhysicsOnly = (bool)MOP.satsumaTogglePhysicsOnly.GetValue();
-
+            
+            // Occlusion Culling
             if (viewDistance < minOcclusionDistance)
             {
                 MOP.occlusionDistance.Value = 400;
                 MOP.minOcclusionDistance.Value = 50;
 
-                if (!warningShowed)
+                if (!occlusionDistanceWarningShowed)
                 {
                     MSCLoader.ModUI.ShowMessage("Occlusion Distance cannot be lower than Minimum Occlusion Distance." +
                         "\n\nIf you don't change it, both values will be reset to default!", "Error");
 
-                    warningShowed = true;
+                    occlusionDistanceWarningShowed = true;
                 }
             }
 
@@ -83,18 +98,22 @@
             occlusionSamples = GetOcclusionSamples();
             viewDistance = int.Parse(MOP.occlusionDistance.GetValue().ToString());
             minOcclusionDistance = int.Parse(MOP.minOcclusionDistance.GetValue().ToString());
-
             occlusionMethod = GetOcclusionMethod();
         }
 
+        /// <summary>
+        /// Returns the value that is used to multiplify the active distance of an object.
+        /// So for example, if the default active distance of the object is 200 units, 
+        /// and the multiplication value is 0.5, the actual active distance will be 100 units.
+        /// </summary>
         static float GetActiveDistanceMultiplicationValue()
         {
             switch (ActiveDistance)
             {
-                default:
-                    return 1;
                 case 0:
                     return 0.5f;
+                default: // 1
+                    return 1;
                 case 2:
                     return 2;
                 case 3:
@@ -102,19 +121,19 @@
             }
         }
 
+        /// <summary>
+        /// Returns the occlusion method used
+        /// </summary>
         static OcclusionMethods GetOcclusionMethod()
         {
-            if ((bool)MOP.occlusionDouble.GetValue())
-                return OcclusionMethods.Double;
-
-            return OcclusionMethods.Chequered;
+            return (bool)MOP.occlusionDouble.GetValue() ? OcclusionMethods.Double : OcclusionMethods.Chequered;
         }
 
+        /// <summary>
+        /// Returns the the occlusion sampling value
+        /// </summary>
         static int GetOcclusionSamples()
         {
-            if ((bool)MOP.occlusionSamplesLowest.GetValue())
-                return 10;
-
             if ((bool)MOP.occlusionSamplesLower.GetValue())
                 return 30;
 
@@ -125,6 +144,23 @@
                 return 240;
 
             return 120;
+        }
+        
+        /// <summary>
+        /// Toggles on overridePhysicsToggling.
+        /// </summary>
+        public static void DisablePhysicsToggling()
+        {
+            MSCLoader.ModConsole.Print("With dedication to James ;)");
+            overridePhysicsToggling = true;
+        }
+
+        /// <summary>
+        /// Toggles off overridePhysicsToggling
+        /// </summary>
+        public static void ResetPhysicsToggling()
+        {
+            overridePhysicsToggling = false;
         }
     }
 }